# apps/dashboard/server/src/openapi.yaml
openapi: 3.0.3
info:
  title: WarrantDB API
  version: 1.0.0
...
servers:
  - url: http://localhost:8080/api
paths:
  /cases:
    get:
      summary: List cases
      tags: [Cases]
      parameters:
        - in: query
          name: query
          schema: { type: string }
        - in: query
          name: county
          schema: { type: string, description: "harris|galveston|brazoria|fortbend|jefferson" }
        - in: query
          name: status
          schema: { type: string }
        - in: query
          name: startDate
          schema: { type: string, format: date }
        - in: query
          name: endDate
          schema: { type: string, format: date }
        - in: query
          name: window
          schema: { type: string, enum: [24h,48h,72h] }
        - in: query
          name: sortBy
          schema: { type: string, enum: [booking_date,bond_amount] }
        - in: query
          name: order
          schema: { type: string, enum: [asc,desc], default: desc }
        - in: query
          name: limit
          schema: { type: integer, minimum: 1, maximum: 200, default: 25 }
      responses:
        '200':
          description: Case list
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items: { $ref: '#/components/schemas/Case' }
                  count:
                    type: integer
                  total:
                    type: integer
...
  /users:
    get:
      summary: List users
      tags: [Users]
      parameters:
        - in: query
          name: role
          schema: { type: string }
        - in: query
          name: status
          schema: { type: string, enum: ['active','suspended','invited','pending_mfa','deleted'] }
        - in: query
          name: search
          schema: { type: string, description: 'Match email or displayName (case-insensitive)' }
      responses:
        '200':
          description: User list
          content:
            application/json:
              schema:
                type: object
                properties:
                  users:
                    type: array
                    items:
                      $ref: '#/components/schemas/UserAccount'
        '403':
          description: Forbidden
    post:
      summary: Create or invite a user
      tags: [Users]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserCreateRequest'
      responses:
        '201':
          description: User created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserCreateResponse'
        '200':
          description: User already existed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserCreateResponse'
        '400':
          description: Validation error
        '403':
          description: Forbidden
...
  /payments:
    get:
      summary: List payment transactions
      tags: [Payments]
      parameters:
        - in: query
          name: status
          schema:
            type: string
            description: Filter by payment status
        - in: query
          name: method
          schema:
            type: string
            description: Filter by payment method (card, ach_debit, wire, check)
        - in: query
          name: search
          schema:
            type: string
            description: Match transaction ID, client name, or bond number (case-insensitive)
      responses:
        '200':
          description: Payments list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentListResponse'
        '403':
          description: Forbidden
    post:
      summary: Create a payment (placeholder until gateway integration)
      tags: [Payments]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PaymentCreateRequest'
      responses:
        '202':
          description: Payment accepted for processing
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentCreateResponse'
        '400':
          description: Validation error
        '403':
          description: Forbidden
...
  /payments/stripe/webhook:
    post:
      summary: Stripe webhook callback (raw payload)
      tags: [Payments]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              description: Event payload delivered by Stripe
      responses:
        '200':
          description: Acknowledged
        '400':
          description: Signature verification failed
        '500':
          description: Internal processing error
  /access-requests:
    get:
      summary: List pending access requests
      tags: [Users]
      parameters:
        - in: query
          name: status
          schema: { type: string, enum: ['pending','reviewed','completed','rejected'] }
      responses:
        '200':
          description: Access requests
          content:
            application/json:
              schema:
                type: object
                properties:
                  requests:
                    type: array
                    items:
                      $ref: '#/components/schemas/AccessRequest'
        '403':
          description: Forbidden
...
  /checkins:
    get:
      summary: List check-ins
      tags: [Checkins]
      operationId: listCheckIns
      parameters:
        - in: query
          name: scope
          schema:
            type: string
            enum: [today, upcoming, overdue, all]
            default: today
        - in: query
          name: officer
          schema:
            type: string
            description: Officer id or name fragment
        - in: query
          name: search
          schema:
            type: string
            description: Client name search string
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            maximum: 200
            default: 50
      responses:
        '200':
          description: Paginated check-in list with summary stats
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CheckInCollectionResponse'
              examples:
                sample:
                  value:
                    scope: today
                    limit: 50
                    items:
                      - id: "65f1c2ab9e7f3a0012ab34cd"
                        person: "John Doe"
                        clientName: "John Doe"
                        county: "harris"
                        dueAt: "2025-09-28T09:00:00.000Z"
                        timezone: "America/Chicago"
                        method: "sms"
                        status: "upcoming"
                        contactCount: 1
                        lastContactAt: null
                        remindersEnabled: true
                        gpsEnabled: false
                        pingsPerDay: 0
                        createdAt: "2025-09-27T18:00:00.000Z"
                        updatedAt: "2025-09-27T18:00:00.000Z"
                        attendance: null
                    stats:
                      totalToday: 8
                      overdue: 2
                      completed: 5
                      gpsEnabled: 3
    post:
      summary: Create a check-in
      tags: [Checkins]
      operationId: createCheckIn
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CheckInCreateRequest'
            examples:
              sample:
                value:
                  person: "John Doe"
                  county: "harris"
                  dueAt: "2025-09-28T09:00:00.000Z"
                  timezone: "America/Chicago"
                  officerId: "65f1c2ab9e7f3a0012ab34aa"
                  method: "sms"
                  remindersEnabled: true
                  gpsEnabled: false
                  pingsPerDay: 3
      responses:
        '201':
          description: Newly created check-in
          content:
            application/json:
              schema:
                type: object
                properties:
                  checkIn:
                    $ref: '#/components/schemas/CheckIn'
              examples:
                created:
                  value:
                    checkIn:
                      id: "65f1c2ab9e7f3a0012ab34cd"
                      person: "John Doe"
                      county: "harris"
                      dueAt: "2025-09-28T09:00:00.000Z"
                      timezone: "America/Chicago"
                      method: "sms"
                      status: "pending"
                      remindersEnabled: true
                      gpsEnabled: false
                      pingsPerDay: 3
...
  /messages:
    get:
      summary: List messages
      description: Returns recent inbound and outbound messages with optional case filter.
      tags: [Messages]
      parameters:
        - in: query
          name: caseId
          schema:
            type: string
            description: Filter messages by case identifier.
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            maximum: 200
            default: 50
      responses:
        '200':
          description: Message list
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/Message'
...
components:
  schemas:
    CrmAddress:
      type: object
      properties:
        streetLine1: { type: string }
        streetLine2: { type: string }
        city: { type: string }
        stateCode: { type: string }
        postalCode: { type: string }
        countryCode: { type: string }
    ChecklistItem:
      type: object
      properties:
        key: { type: string }
        label: { type: string }
        required: { type: boolean }
        status: { type: string, enum: [pending, completed] }
        completedAt: { type: string, format: date-time }
        note: { type: string }
    Case:
      type: object
      properties:
        id: { type: string }
        full_name: { type: string }
        county: { type: string }
        booking_date: { type: string }
        bond_amount: { type: number, nullable: true }
        bond_label: { type: string }
        case_number: { type: string }
        spn: { type: string }
        phone: { type: string }
        address: { $ref: '#/components/schemas/CaseSourceAddress' }
        crm_stage: { type: string, enum: [new, contacted, qualifying, accepted, denied] }
        crm_details:
          type: object
          properties:
            assignedTo: { type: string }
            followUpAt: { type: string, format: date-time }
            phone: { type: string }
            address: { $ref: '#/components/schemas/CrmAddress' }
            documents:
              type: array
              items: { $ref: '#/components/schemas/ChecklistItem' }
            attachments:
              type: array
              items:
                type: object
                additionalProperties: true
    Message:
      type: object
      properties:
        id: { type: string }
        caseId: { type: string, nullable: true }
        direction: { type: string, enum: [out, in] }
        channel: { type: string, enum: [sms, voice] }
        status: { type: string, enum: [queued, sending, sent, delivered, failed] }
        to: { type: string, nullable: true }
        from: { type: string, nullable: true }
        body: { type: string, nullable: true }
        provider: { type: string, nullable: true }
        providerMessageId: { type: string, nullable: true }
        createdAt: { type: string, format: date-time, nullable: true }
        sentAt: { type: string, format: date-time, nullable: true }
        deliveredAt: { type: string, format: date-time, nullable: true }
        errorCode: { type: string, nullable: true }
        errorMessage: { type: string, nullable: true }
        meta:
          type: object
          additionalProperties: true
    CaseSourceAddress:
      type: object
      properties:
        line1: { type: string }
        line2: { type: string }
        city: { type: string }
        state: { type: string }
        zip: { type: string }
        postalCode: { type: string }
...
    Money:
      type: number
      format: double
      example: 15000
...
    UserAccount:
      type: object
      properties:
        uid: { type: string }
        email: { type: string, format: email }
        displayName: { type: string, nullable: true }
        roles:
          type: array
          items: { type: string }
        departments:
          type: array
          items: { type: string }
        counties:
          type: array
          items: { type: string }
        status:
          type: string
          enum: ['active','suspended','invited','pending_mfa','deleted']
        mfaEnforced: { type: boolean }
        lastLoginAt: { type: string, format: date-time, nullable: true }
        invitedAt: { type: string, format: date-time, nullable: true }
        invitedBy: { type: string, nullable: true }
        lastRoleChangeAt: { type: string, format: date-time, nullable: true }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    UserCreateRequest:
      type: object
      required: [email]
      properties:
        email: { type: string, format: email }
        displayName: { type: string }
        roles:
          type: array
          items: { type: string }
        departments:
          type: array
          items: { type: string }
        counties:
          type: array
          items: { type: string }
        status:
          type: string
          enum: ['active','suspended','invited','pending_mfa','deleted']
    UserCreateResponse:
      type: object
      properties:
        user: { $ref: '#/components/schemas/UserAccount' }
        inviteLink:
          type: string
          nullable: true
          description: Password setup link that can be shared securely
...
    Payment:
      type: object
      properties:
        id: { type: string }
        transactionId: { type: string }
        amount: { $ref: '#/components/schemas/Money' }
        fees: { $ref: '#/components/schemas/Money', nullable: true }
        netAmount: { $ref: '#/components/schemas/Money', nullable: true }
        currency: { type: string, example: 'usd' }
        method: { type: string, example: 'card' }
        status: { type: string, example: 'completed' }
        bondNumber: { type: string, nullable: true }
        clientName: { type: string, nullable: true }
        clientEmail: { type: string, format: email, nullable: true }
        processedAt: { type: string, format: date-time, nullable: true }
        createdAt: { type: string, format: date-time, nullable: true }
        updatedAt: { type: string, format: date-time, nullable: true }
        flags:
          type: array
          items: { type: string }
        failureReason: { type: string, nullable: true }
        processor:
          type: object
          additionalProperties: true
        stripePaymentIntentId: { type: string, nullable: true }
        stripeChargeId: { type: string, nullable: true }
        createdByUid: { type: string, nullable: true }
    PaymentListResponse:
      type: object
      properties:
        items:
          type: array
          items: { $ref: '#/components/schemas/Payment' }
        total: { type: integer }
    PaymentCreateRequest:
      type: object
      required: [amount]
      properties:
        amount: { $ref: '#/components/schemas/Money' }
        currency: { type: string, default: 'usd' }
        method: { type: string, default: 'card' }
        clientName: { type: string }
        clientEmail: { type: string, format: email }
        bondNumber: { type: string }
        metadata:
          type: object
          additionalProperties: true
    PaymentCreateResponse:
      type: object
      properties:
        payment:
          $ref: '#/components/schemas/Payment'
        clientSecret:
          type: string
          description: Stripe payment intent client secret (use with Stripe.js)
...
    AccessRequest:
      type: object
      properties:
        id: { type: string }
        email: { type: string, format: email }
        displayName: { type: string, nullable: true }
        message: { type: string, nullable: true }
        status: { type: string, enum: ['pending','reviewed','completed','rejected'] }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
...
    CheckIn:
      type: object
      properties:
        id:
          type: string
        clientId:
          type: string
          nullable: true
        caseId:
          type: string
          nullable: true
        clientName:
          type: string
        person:
          type: string
        caseNumber:
          type: string
          nullable: true
        county:
          type: string
          nullable: true
        dueAt:
          type: string
          format: date-time
        timezone:
          type: string
        officerId:
          type: string
          nullable: true
        method:
          type: string
          enum: [sms, phone, call, in-person, app]
        status:
          type: string
          enum: [pending, upcoming, overdue, done]
        note:
          type: string
          nullable: true
        contactCount:
          type: integer
        lastContactAt:
          type: string
          format: date-time
          nullable: true
        location:
          $ref: '#/components/schemas/CheckInLocation'
        remindersEnabled:
          type: boolean
        gpsEnabled:
          type: boolean
        pingsPerDay:
          type: integer
        lastPingAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        attendance:
          $ref: '#/components/schemas/CheckInAttendance'
    CheckInLocation:
      type: object
      nullable: true
      properties:
        lat:
          type: number
        lng:
          type: number
        accuracy:
          type: number
    CheckInAttendance:
      type: object
      nullable: true
      properties:
        status:
          type: string
          enum: [attended, missed]
        recordedAt:
          type: string
          format: date-time
          nullable: true
        recordedBy:
          type: string
          nullable: true
        note:
          type: string
          nullable: true
        location:
          $ref: '#/components/schemas/CheckInLocation'
    CheckInStats:
      type: object
      properties:
        totalToday:
          type: integer
        overdue:
          type: integer
        completed:
          type: integer
        gpsEnabled:
          type: integer
    CheckInCollectionResponse:
      type: object
      properties:
        scope:
          type: string
        limit:
          type: integer
        items:
          type: array
          items:
            $ref: '#/components/schemas/CheckIn'
        stats:
          $ref: '#/components/schemas/CheckInStats'
...
    CheckInCreateRequest:
      type: object
      required: [dueAt, person]
      properties:
        clientId:
          type: string
        caseId:
          type: string
        person:
          type: string
        county:
          type: string
        dueAt:
          type: string
          format: date-time
        timezone:
          type: string
        officerId:
          type: string
        method:
          type: string
          enum: [sms, phone, call, in-person, app]
        notes:
          type: string
        remindersEnabled:
          type: boolean
        gpsEnabled:
          type: boolean
        pingsPerDay:
          type: integer
          minimum: 1
          maximum: 12
        locationText:
          type: string
        caseNumber:
          type: string
          description: Optional reference case number for downstream reporting
        personName:
          type: string
          description: Alias for person when supplied by the client selector
...